% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaMatrixExample}
\alias{armaMatrixExample}
\title{Given a (nano)arrow object via two pointers, and a column size, return an arma matrix}
\usage{
armaMatrixExample(vec, ncol, verbose = FALSE)
}
\arguments{
\item{vec}{A nanoarrow object}

\item{ncol}{Number of columns in returned matrix}

\item{verbose}{A logical value, default is false}
}
\description{
Given a (nano)arrow object via two pointers, and a column size, return an arma matrix
}
\section{Limitations}{

As this aims at \pkg{armadillo} operations, this functionality is limited to numeric
vector columns. The framework used here could of course be extended to other Arrow formats.
}

\examples{
sv <- nanoarrow::as_nanoarrow_array(c(4:8), nanoarrow::na_int16())  # internally int16
armaMatrixExample(sv, 2)
}
