% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaVectorExample}
\alias{armaVectorExample}
\title{Given a (nano)arrow array external pointer object via two pointers (with the second tucked
away as the tag), return an arma vector}
\usage{
armaVectorExample(vec, verbose = FALSE)
}
\arguments{
\item{vec}{A nanoarrow array object as an external pointer}

\item{verbose}{A logical value, default is false}
}
\description{
Given a (nano)arrow array external pointer object via two pointers (with the second tucked
away as the tag), return an arma vector
}
\section{Limitations}{

As this aims at \pkg{armadillo} interoperation, this functionality is limited to numeric
vector columns. The framework used here could of course be extended to other Arrow formats.
}

\examples{
sv <- nanoarrow::as_nanoarrow_array(c(4:7), nanoarrow::na_int16())  # internally int16
armaVectorExample(sv)
}
