% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{structExample}
\alias{structExample}
\title{Given a vector of JSON expressions and a vector of format strings, return a struct}
\usage{
structExample(jsontxt, format)
}
\arguments{
\item{jsontxt}{A vector of JSON strings (as understood by \pkg{nanoarrow} testing helper code)s}

\item{format}{A vector of format strings per the Arrow specification}
}
\description{
Given a vector of JSON expressions and a vector of format strings, return a struct
}
\section{Limitations}{

As this aims at \pkg{armadillo} interoperation, thie functionality is limited to numeric
vector columns. The framework used here could of course be extended to other Arrow formats.
}

\examples{
structExample(rep(R"({"name": "vec", "count": 3, "VALIDITY": [1, 1, 1], "DATA": ["2", "3", "4"]})",2), c("S", "L"))
# also pipe above into either `tibble::as_tibble()` (converts data) or `arrow::as_record_batch()`
}
