% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectorExample}
\alias{vectorExample}
\title{Given a JSON expressions and a string, return a vector}
\usage{
vectorExample(jsontxt, format)
}
\arguments{
\item{jsontxt}{A JSON string (as understood by \pkg{nanoarrow} testing helper code)s}

\item{format}{A format string per the Arrow specification}
}
\description{
Given a JSON expressions and a string, return a vector
}
\section{Limitations}{

As this aims at \pkg{armadillo} operation, this functionality is limited to numeric
vector columns. The framework used here could of course be extended to other Arrow formats.
}

\examples{
vectorExample(R"({"name": "vec", "count": 3, "VALIDITY": [1, 1, 1], "DATA": ["2", "3", "4"]})", "S")
}
