% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaMatrixExample}
\alias{armaMatrixExample}
\title{Given a (nano)arrow object via two pointers, and a column size, return an arma matrix}
\usage{
armaMatrixExample(vec, ncol, verbose = FALSE)
}
\arguments{
\item{vec}{A nanoarrow object}

\item{verbose}{A logical value, default is false}
}
\description{
Given a (nano)arrow object via two pointers, and a column size, return an arma matrix
}
\section{Limitations}{

As this aims at \pkg{armadillo} operations, this functionality is limited to numeric
vector columns. The framework used here could of course be extended to other Arrow formats.
}

\examples{
sv <- vectorExample(R"({"name": "vec", "count": 4, "VALIDITY": [1, 1, 1, 1], "DATA": [2, 3, 4, 5]})", "g")
armaMatrixExample(sv, 2)
}
