% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaVectorExample}
\alias{armaVectorExample}
\title{Given a (nano)arrow object via two pointers, return an arma vector}
\usage{
armaVectorExample(vec, verbose = FALSE)
}
\arguments{
\item{vec}{A nanoarrow object}

\item{verbose}{A logical value, default is false}
}
\description{
Given a (nano)arrow object via two pointers, return an arma vector
}
\section{Limitations}{

As this aims at \pkg{armadillo} interoperation, thie functionality is limited to numeric
vector columns. The framework used here could of course be extended to other Arrow formats.
}

\examples{
sv <- vectorExample(R"({"name": "vec", "count": 3, "VALIDITY": [1, 1, 1], "DATA": [2, 3, 4]})", "g")
armaVectorExample(sv)
}
